chore: initial split of gesture editor into multiple files
