--- app/common/renderer/components/SessionInspector/GesturesTab/GestureEditor.jsx
@@ -1,727 +1,0 @@
-import {
-  IconArrowBigDownLine,
-  IconArrowBigRightLines,
-  IconArrowBigUpLine,
-  IconArrowLeft,
-  IconFocus2,
-  IconHandMove,
-  IconPlayerPause,
-  IconPlayerPlay,
-  IconPlus,
-  IconQuestionMark,
-  IconX,
-} from '@tabler/icons-react';
-import {
-  Button,
-  Card,
-  Col,
-  Divider,
-  Flex,
-  Input,
-  Popover,
-  Row,
-  Select,
-  Space,
-  Steps,
-  Tabs,
-  Tooltip,
-} from 'antd';
-import _ from 'lodash';
-import {useCallback, useEffect, useState} from 'react';
-import {useTranslation} from 'react-i18next';
-
-import {NOTIF, TABLE_TAB} from '../../../constants/antd-types.js';
-import {
-  CURSOR,
-  DEFAULT_POINTER,
-  POINTER_COLORS,
-  POINTER_DOWN_BTNS,
-  POINTER_MOVE_COORDS_TYPE,
-  POINTER_MOVE_DEFAULT_DURATION,
-  POINTER_TYPES,
-  POINTER_TYPES_MAP,
-  TICK_PROPS,
-} from '../../../constants/gestures.js';
-import {SCREENSHOT_INTERACTION_MODE} from '../../../constants/screenshot.js';
-import {notification} from '../../../utils/notification.js';
-import {percentageToPixels, pixelsToPercentage} from '../../../utils/other.js';
-import inspectorStyles from '../SessionInspector.module.css';
-import styles from './Gestures.module.css';
-
-const {POINTER_UP, POINTER_DOWN, PAUSE, POINTER_MOVE, FILLER} = POINTER_TYPES;
-
-/**
- * Shows the gesture editor interface
- */
-const GestureEditor = (props) => {
-  const {
-    loadedGesture,
-    saveGesture,
-    tickCoordinates,
-    selectedTick,
-    selectTick,
-    unselectTick,
-    windowSize,
-    displayGesture,
-  } = props;
-  const {t} = useTranslation();
-
-  const defaultGestureName = t('Untitled Gesture');
-  const defaultGestureDescription = t('Add Description');
-
-  const [pointers, setPointers] = useState(loadedGesture?.actions ?? DEFAULT_POINTER);
-  const [name, setName] = useState(loadedGesture?.name ?? defaultGestureName);
-  const [description, setDescription] = useState(
-    loadedGesture?.description ?? defaultGestureDescription,
-  );
-  const [coordType, setCoordType] = useState(POINTER_MOVE_COORDS_TYPE.PERCENTAGES);
-  const [activePointerId, setActivePointerId] = useState('1');
-
-  const onSave = () => {
-    const {id, date} = loadedGesture;
-    if (duplicatePointerNames(pointers)) {
-      return null;
-    }
-    const gesture = {
-      name,
-      description,
-      id,
-      date,
-      actions: getConvertedPointers(POINTER_MOVE_COORDS_TYPE.PERCENTAGES),
-    };
-    saveGesture(gesture);
-    displayNotificationMsg(NOTIF.SUCCESS, t('Gesture saved'));
-  };
-
-  const onSaveAs = () => {
-    if (duplicatePointerNames(pointers)) {
-      return null;
-    }
-    const gesture = {
-      name,
-      description,
-      actions: getConvertedPointers(POINTER_MOVE_COORDS_TYPE.PERCENTAGES),
-    };
-    saveGesture(gesture);
-    displayNotificationMsg(NOTIF.SUCCESS, t('Gesture saved as', {gestureName: name}));
-    onBack();
-  };
-
-  const onPlay = () => {
-    const {applyClientMethod} = props;
-    if (duplicatePointerNames(pointers)) {
-      return null;
-    }
-    const formattedPointers = getW3CPointers();
-    applyClientMethod({methodName: SCREENSHOT_INTERACTION_MODE.GESTURE, args: [formattedPointers]});
-  };
-
-  const onBack = () => {
-    const {hideGestureEditor, removeLoadedGesture, removeGestureDisplay} = props;
-    unselectTick();
-    removeGestureDisplay();
-    removeLoadedGesture();
-    hideGestureEditor();
-  };
-
-  // Check if pointer names are duplicates before saving/playing
-  const duplicatePointerNames = (localPointers) => {
-    const duplicates = {};
-    for (const pointer of localPointers) {
-      if (duplicates[pointer.name]) {
-        displayNotificationMsg(NOTIF.ERROR, t('Duplicate pointer names are not allowed'));
-        return true;
-      } else {
-        duplicates[pointer.name] = pointer;
-      }
-    }
-    return false;
-  };
-
-  const displayNotificationMsg = (type, msg) => {
-    notification[type]({
-      title: msg,
-      duration: 5,
-    });
-  };
-
-  // Change gesture datastructure to fit Webdriver spec
-  const getW3CPointers = () => {
-    const newPointers = {};
-    const currentPointers = getConvertedPointers(POINTER_MOVE_COORDS_TYPE.PIXELS);
-    for (const pointer of currentPointers) {
-      newPointers[pointer.name] = pointer.ticks.map((tick) => _.omit(tick, 'id'));
-    }
-    return newPointers;
-  };
-
-  // This converts all the coordinates in the gesture to px/%
-  const getConvertedPointers = useCallback(
-    (type) => {
-      const {width, height} = windowSize;
-      if (type === coordType) {
-        return pointers;
-      }
-      const newPointers = _.cloneDeep(pointers);
-      for (const pointer of newPointers) {
-        for (const tick of pointer.ticks) {
-          if (tick.type === POINTER_TYPES.POINTER_MOVE) {
-            if (type === POINTER_MOVE_COORDS_TYPE.PIXELS) {
-              tick.x = percentageToPixels(tick.x, width);
-              tick.y = percentageToPixels(tick.y, height);
-            } else {
-              tick.x = pixelsToPercentage(tick.x, width);
-              tick.y = pixelsToPercentage(tick.y, height);
-            }
-          }
-        }
-      }
-      return newPointers;
-    },
-    [coordType, pointers, windowSize],
-  );
-
-  const getDefaultMoveDuration = useCallback(
-    (ticks, tickId, x2, y2, coordFromTap) => {
-      const {width, height} = windowSize;
-      const ticksExceptCurrent = ticks.filter((tick) => tick.id !== tickId);
-      const prevPointerMoves = [];
-      for (const tick of ticksExceptCurrent) {
-        if (tick.type === POINTER_MOVE && tick.x !== undefined && tick.y !== undefined) {
-          prevPointerMoves.push({x: tick.x, y: tick.y});
-        }
-      }
-      const len = prevPointerMoves.length;
-      if (len === 0) {
-        return 0;
-      }
-      const obj = {x1: prevPointerMoves[len - 1].x, y1: prevPointerMoves[len - 1].y, x2, y2};
-      if (coordType === POINTER_MOVE_COORDS_TYPE.PERCENTAGES) {
-        obj.x1 = percentageToPixels(obj.x1, width);
-        obj.y1 = percentageToPixels(obj.y1, height);
-        // No need to convert coordinates from tap since they are in px
-        if (!coordFromTap) {
-          obj.x2 = percentageToPixels(obj.x2, width);
-          obj.y2 = percentageToPixels(obj.y2, height);
-        }
-      }
-      const calcLength = (v1, v2) => Math.sqrt(v1 ** 2 + v2 ** 2);
-      const calcDiff = (v1, v2) => Math.abs(v2) - Math.abs(v1);
-      const xDiff = calcDiff(obj.x1, obj.x2);
-      const yDiff = calcDiff(obj.y1, obj.y2);
-      const maxScreenLength = calcLength(width, height);
-      const lineLength = calcLength(xDiff, yDiff);
-      const lineLengthPct = lineLength / maxScreenLength;
-      return Math.round(lineLengthPct * POINTER_MOVE_DEFAULT_DURATION);
-    },
-    [coordType, windowSize],
-  );
-
-  // Update tapped coordinates within local state
-  const updateCoordinates = useCallback(
-    (tickKey, updateX, updateY) => {
-      if (!updateX || !updateY) {
-        return null;
-      }
-      const {width, height} = windowSize;
-      const copiedPointers = _.cloneDeep(pointers);
-      const currentPointer = copiedPointers.find((pointer) => pointer.id === tickKey[0]);
-      const currentTick = currentPointer.ticks.find((tick) => tick.id === tickKey);
-      const x = parseFloat(updateX);
-      const y = parseFloat(updateY);
-      if (coordType === POINTER_MOVE_COORDS_TYPE.PERCENTAGES) {
-        currentTick.x = pixelsToPercentage(x, width);
-        currentTick.y = pixelsToPercentage(y, height);
-      } else {
-        currentTick.x = x;
-        currentTick.y = y;
-      }
-
-      if (currentTick.duration === undefined) {
-        currentTick.duration = getDefaultMoveDuration(
-          currentPointer.ticks,
-          currentTick.id,
-          x,
-          y,
-          true,
-        );
-      }
-      setPointers(copiedPointers);
-    },
-    [coordType, getDefaultMoveDuration, pointers, windowSize],
-  );
-
-  const addPointer = () => {
-    const key = pointers.length + 1;
-    const pointerId = String(key);
-    const copiedPointers = _.cloneDeep(pointers);
-    copiedPointers.push({
-      name: `pointer${key}`,
-      ticks: [{id: `${key}.1`}],
-      id: pointerId,
-      color: POINTER_COLORS[key - 1],
-    });
-    setPointers(copiedPointers);
-    setActivePointerId(pointerId);
-  };
-
-  const deletePointer = (targetKey) => {
-    // 'newActivePointerId' variable keeps track of the previous pointer before deleting the current one
-    // its default is the first pointer
-    let newActivePointerId = '1';
-    const pointersExceptCurrent = pointers.filter((pointer) => pointer.id !== targetKey);
-    const newPointers = pointersExceptCurrent.map((pointer, index) => {
-      const id = String(index + 1);
-      if (id !== pointer.id) {
-        pointer.id = id;
-        pointer.color = POINTER_COLORS[index];
-        pointer.ticks = pointer.ticks.map((tick) => {
-          tick.id = `${id}.${tick.id[2]}`;
-          return tick;
-        });
-      } else {
-        newActivePointerId = pointer.id;
-      }
-      return pointer;
-    });
-    unselectTick();
-    setPointers(newPointers);
-    setActivePointerId(newActivePointerId);
-  };
-
-  const addTick = (pointerKey) => {
-    const copiedPointers = _.cloneDeep(pointers);
-    const currentPointer = copiedPointers.find((pointer) => pointer.id === pointerKey);
-    const id = `${pointerKey}.${currentPointer.ticks.length + 1}`;
-    currentPointer.ticks.push({id});
-    setPointers(copiedPointers);
-  };
-
-  const deleteTick = (pointerKey, tickKey) => {
-    const copiedPointers = _.cloneDeep(pointers);
-    const currentPointer = copiedPointers.find((pointer) => pointer.id === pointerKey);
-    const ticksToKeep = currentPointer.ticks.filter((tick) => tick.id !== tickKey);
-    const newTicks = ticksToKeep.map((tick, index) => {
-      const id = String(index + 1);
-      if (tick.id !== id) {
-        tick.id = `${tick.id[0]}.${id}`;
-      }
-      return tick;
-    });
-    currentPointer.ticks = newTicks;
-    unselectTick();
-    setPointers(copiedPointers);
-  };
-
-  // Updates the current tick within local state
-  const updateTick = (tick, msg, value) => {
-    const copiedPointers = _.cloneDeep(pointers);
-    const currentPointer = copiedPointers.find((p) => p.id === tick.id[0]);
-    const targetTickIdx = currentPointer.ticks.findIndex((t) => t.id === tick.id);
-    // currentTick can be assigned a new tick object if made changes to pointer types
-    let currentTick = currentPointer.ticks[targetTickIdx];
-
-    // We create an entire new tick for changes in pointer types to ensure previous properties are removed
-    if (msg === TICK_PROPS.POINTER_TYPE) {
-      if (value === POINTER_MOVE) {
-        selectTick(tick.id);
-      }
-      currentTick = {
-        id: tick.id,
-        type: value,
-        ...([POINTER_DOWN, POINTER_UP].includes(value) && {button: POINTER_DOWN_BTNS.LEFT}),
-        ...(value === PAUSE && {duration: 0}),
-      };
-    } else {
-      // We just modify the existing tick values
-      currentTick[msg] = parseFloat(value);
-
-      // set default duration for if not set already
-      if (
-        currentTick.x !== undefined &&
-        currentTick.y !== undefined &&
-        currentTick.duration === undefined
-      ) {
-        currentTick.duration = getDefaultMoveDuration(
-          currentPointer.ticks,
-          tick.id,
-          currentTick.x,
-          currentTick.y,
-          false,
-        );
-      }
-    }
-
-    currentPointer.ticks[targetTickIdx] = currentTick;
-    setPointers(copiedPointers);
-  };
-
-  // Reformats the gesture only for the timeline by populating the 'filler' ticks for each pointer
-  // to match same length to keep timeline lengths consistent and accurate
-  const updateGestureForTimeline = () => {
-    const copiedPointers = _.cloneDeep(pointers);
-    const allTickLengths = copiedPointers.map((pointer) => pointer.ticks.length);
-    const maxTickLength = Math.max(...allTickLengths);
-    return copiedPointers.map((pointer) => {
-      const currentLength = pointer.ticks.length;
-      if (currentLength > 0 && currentLength < maxTickLength) {
-        const fillers = Array.from({length: maxTickLength - currentLength}, () => ({
-          type: FILLER,
-        }));
-        pointer.ticks.push(...fillers);
-      }
-      return pointer;
-    });
-  };
-
-  const updatePointerName = (pointerName, pointerIndex) => {
-    const copiedPointers = _.cloneDeep(pointers);
-    copiedPointers[pointerIndex].name = pointerName;
-    setPointers(copiedPointers);
-  };
-
-  // Draw gesture whenever pointers change
-  useEffect(() => {
-    const convertedPointers = getConvertedPointers(POINTER_MOVE_COORDS_TYPE.PIXELS);
-    displayGesture(convertedPointers);
-  }, [displayGesture, getConvertedPointers, pointers]);
-
-  // Retrieve coordinates when user taps screenshot
-  // Defer state updates to the next animation frame to avoid synchronous setState in effect
-  useEffect(() => {
-    if (!tickCoordinates || !selectedTick) {
-      return;
-    }
-    let rafId = requestAnimationFrame(() => {
-      updateCoordinates(selectedTick, tickCoordinates.x, tickCoordinates.y);
-    });
-    return () => cancelAnimationFrame(rafId);
-  }, [selectedTick, tickCoordinates, updateCoordinates]);
-
-  const headerTitle = (
-    <Tooltip title={t('Edit')}>
-      <Input
-        defaultValue={name}
-        className={styles.gestureHeaderTitle}
-        onChange={(e) => setName(e.target.value)}
-        size="small"
-      />
-    </Tooltip>
-  );
-
-  const headerButtons = (
-    <Space>
-      <Space.Compact>
-        <Tooltip title={t('showMoveActionCoordsInPercentage')}>
-          <Button
-            type={coordType === POINTER_MOVE_COORDS_TYPE.PERCENTAGES ? 'primary' : 'default'}
-            onClick={() => {
-              setPointers(getConvertedPointers(POINTER_MOVE_COORDS_TYPE.PERCENTAGES));
-              setCoordType(POINTER_MOVE_COORDS_TYPE.PERCENTAGES);
-            }}
-          >
-            %
-          </Button>
-        </Tooltip>
-        <Tooltip title={t('showMoveActionCoordsInPixels')}>
-          <Button
-            type={coordType === POINTER_MOVE_COORDS_TYPE.PIXELS ? 'primary' : 'default'}
-            onClick={() => {
-              setPointers(getConvertedPointers(POINTER_MOVE_COORDS_TYPE.PIXELS));
-              setCoordType(POINTER_MOVE_COORDS_TYPE.PIXELS);
-            }}
-          >
-            px
-          </Button>
-        </Tooltip>
-      </Space.Compact>
-      <Tooltip title={t('Play')}>
-        <Button type="primary" icon={<IconPlayerPlay size={18} />} onClick={() => onPlay()} />
-      </Tooltip>
-      <Space.Compact>
-        <Button onClick={() => onSaveAs()}>{t('saveAs')}</Button>
-        <Button onClick={() => onSave()} disabled={!loadedGesture}>
-          {t('Save')}
-        </Button>
-      </Space.Compact>
-    </Space>
-  );
-
-  const headerDescription = (
-    <div className={styles.gestureHeaderDescription}>
-      <Tooltip title={t('Edit')}>
-        <Input
-          defaultValue={description}
-          className={styles.gestureHeaderDescriptionInput}
-          onChange={(e) => setDescription(e.target.value)}
-          size="small"
-        />
-      </Tooltip>
-    </div>
-  );
-
-  const header = (
-    <>
-      <Row justify="space-between">
-        <Col flex="32px">
-          <Button type="text" icon={<IconArrowLeft size={18} />} onClick={() => onBack()} />
-        </Col>
-        <Col flex="1">{headerTitle}</Col>
-        <Col>{headerButtons}</Col>
-      </Row>
-      {headerDescription}
-    </>
-  );
-
-  const regularTimelineIcon = (tick) => {
-    const {type, duration, button, x, y} = tick;
-    return (
-      <Popover
-        placement="bottom"
-        title={<center>{t(POINTER_TYPES_MAP[type])}</center>}
-        content={
-          <div className={styles.timelineTickTitle}>
-            {duration !== undefined && (
-              <p>
-                {t('Duration')}: {duration}ms
-              </p>
-            )}
-            {button !== undefined && (
-              <p>
-                {t('Button')}: {button === POINTER_DOWN_BTNS.LEFT ? t('Left') : t('Right')}
-              </p>
-            )}
-            {x !== undefined && (
-              <p>
-                X: {x}
-                {coordType === POINTER_MOVE_COORDS_TYPE.PIXELS ? 'px' : '%'}
-              </p>
-            )}
-            {y !== undefined && (
-              <p>
-                Y: {y}
-                {coordType === POINTER_MOVE_COORDS_TYPE.PIXELS ? 'px' : '%'}
-              </p>
-            )}
-            {type === undefined && <p>{t('Action Type Not Defined')}</p>}
-          </div>
-        }
-      >
-        {type === POINTER_MOVE && <IconArrowBigRightLines size={18} />}
-        {type === POINTER_DOWN && <IconArrowBigDownLine size={18} />}
-        {type === POINTER_UP && <IconArrowBigUpLine size={18} />}
-        {type === PAUSE && <IconPlayerPause size={18} />}
-        {type === undefined && <IconQuestionMark size={18} />}
-      </Popover>
-    );
-  };
-
-  const timeline = updateGestureForTimeline().map((pointer) => (
-    <center key={pointer.id}>
-      <Steps
-        className={styles.gestureTimeline}
-        style={{'--timelineColor': pointer.color}}
-        items={pointer.ticks.map((tick) =>
-          tick.type !== FILLER
-            ? {
-                status: 'finish',
-                icon: regularTimelineIcon(tick),
-              }
-            : {
-                status: 'wait',
-                icon: <IconQuestionMark size={18} />,
-              },
-        )}
-      />
-    </center>
-  ));
-
-  const tickButton = (tick) => (
-    <center>
-      <Space.Compact block>
-        <Button
-          block
-          type={tick.button === POINTER_DOWN_BTNS.LEFT ? 'primary' : 'default'}
-          onClick={() => updateTick(tick, TICK_PROPS.BUTTON, POINTER_DOWN_BTNS.LEFT)}
-        >
-          {t('Left')}
-        </Button>
-        <Button
-          block
-          type={tick.button === POINTER_DOWN_BTNS.RIGHT ? 'primary' : 'default'}
-          onClick={() => updateTick(tick, TICK_PROPS.BUTTON, POINTER_DOWN_BTNS.RIGHT)}
-        >
-          {t('Right')}
-        </Button>
-      </Space.Compact>
-    </center>
-  );
-
-  const tickDuration = (tick) => (
-    <center>
-      <Space.Compact block>
-        <Input
-          className={styles.tickInputBox}
-          value={!isNaN(tick.duration) ? tick.duration : null}
-          placeholder={t('Duration')}
-          defaultValue={tick.duration}
-          onChange={(e) => updateTick(tick, TICK_PROPS.DURATION, e.target.value)}
-        />
-        <Space.Addon>ms</Space.Addon>
-      </Space.Compact>
-    </center>
-  );
-
-  const tickCoords = (tick) => (
-    <center>
-      <Space.Compact block>
-        <Input
-          className={styles.tickInputBox}
-          value={!isNaN(tick.x) ? tick.x : ''}
-          placeholder="X"
-          defaultValue={tick.x}
-          onChange={(e) => updateTick(tick, TICK_PROPS.X, e.target.value)}
-        />
-        <Input
-          className={styles.tickInputBox}
-          value={!isNaN(tick.y) ? tick.y : ''}
-          placeholder="Y"
-          defaultValue={tick.y}
-          onChange={(e) => updateTick(tick, TICK_PROPS.Y, e.target.value)}
-        />
-      </Space.Compact>
-    </center>
-  );
-
-  const tickType = (tick) => (
-    <center>
-      <Select
-        styles={{
-          root: {width: '100%'},
-          content: {justifyContent: 'center'},
-          popup: {listItem: {textAlign: 'center'}},
-        }}
-        placeholder={t('Action Type')}
-        value={tick.type}
-        defaultValue={tick.type}
-        size="middle"
-        popupMatchSelectWidth={false}
-        onChange={(e) => updateTick(tick, TICK_PROPS.POINTER_TYPE, e)}
-        options={[
-          {value: POINTER_MOVE, label: t(POINTER_TYPES_MAP.pointerMove)},
-          {value: POINTER_DOWN, label: t(POINTER_TYPES_MAP.pointerDown)},
-          {value: POINTER_UP, label: t(POINTER_TYPES_MAP.pointerUp)},
-          {value: PAUSE, label: t(POINTER_TYPES_MAP.pause)},
-        ]}
-      />
-    </center>
-  );
-
-  const tapCoordinatesBtn = (tickId) => (
-    <Tooltip title={t('toggleMoveActionCoordPicker')}>
-      <Button
-        size="small"
-        type={selectedTick === tickId ? 'primary' : 'text'}
-        icon={<IconFocus2 size={18} />}
-        onClick={() => (selectedTick === tickId ? unselectTick() : selectTick(tickId))}
-      />
-    </Tooltip>
-  );
-
-  const tickCard = (tick) => (
-    <Card
-      hoverable={true}
-      className={styles.tickCard}
-      extra={
-        <>
-          {tick.type === POINTER_MOVE && tapCoordinatesBtn(tick.id)}
-          <Tooltip title={t('Delete')}>
-            <Button
-              size="small"
-              type="text"
-              icon={<IconX size={18} />}
-              onClick={() => deleteTick(tick.id[0], tick.id)}
-            />
-          </Tooltip>
-        </>
-      }
-    >
-      <Space className={inspectorStyles.spaceContainer} orientation="vertical" size="middle">
-        {tickType(tick)}
-        {(tick.type === POINTER_MOVE || tick.type === PAUSE) && tickDuration(tick)}
-        {(tick.type === POINTER_DOWN || tick.type === POINTER_UP) && tickButton(tick)}
-        {tick.type === POINTER_MOVE && tickCoords(tick)}
-      </Space>
-    </Card>
-  );
-
-  const pointerTicksGrid = (pointer) => (
-    <Row gutter={[24, 24]}>
-      {pointer.ticks.map((tick) => (
-        <Col xs={12} sm={12} md={12} lg={8} xl={6} xxl={4} xxxl={3} key={tick.id}>
-          {tickCard(tick)}
-        </Col>
-      ))}
-      <Col xs={12} sm={12} md={12} lg={8} xl={6} xxl={4} xxxl={3}>
-        <div className={styles.tickPlusBtnWrapper}>
-          <Tooltip title={t('Add')}>
-            <Button icon={<IconPlus size={18} />} onClick={() => addTick(pointer.id)} />
-          </Tooltip>
-        </div>
-      </Col>
-    </Row>
-  );
-
-  const pointerTabs = pointers.map((pointer, index) => ({
-    label: (
-      <Tooltip title={t('Edit')}>
-        <Input
-          className={styles.pointerTitle}
-          style={{
-            cursor: activePointerId === pointer.id ? CURSOR.TEXT : CURSOR.POINTER,
-            textDecorationColor: pointer.color,
-          }}
-          value={pointer.name}
-          defaultValue={pointer.name}
-          variant="borderless"
-          maxLength={10}
-          onChange={(e) => updatePointerName(e.target.value, index)}
-        />
-      </Tooltip>
-    ),
-    key: pointer.id,
-    closable: pointer.id !== '1',
-    children: pointerTicksGrid(pointer),
-  }));
-
-  return (
-    <Card
-      title={
-        <Flex gap={4} align="center">
-          <IconHandMove size={18} />
-          {t('Gesture Builder')}
-        </Flex>
-      }
-      className={inspectorStyles.interactionTabCard}
-    >
-      {header}
-      <Divider className={styles.gestureEditorDivider} />
-      {timeline}
-      <Divider className={styles.gestureEditorDivider} />
-      <Tabs
-        type="editable-card"
-        onChange={(pointerId) => setActivePointerId(pointerId)}
-        activeKey={activePointerId}
-        onEdit={(targetKey, action) =>
-          action === TABLE_TAB.ADD ? addPointer() : deletePointer(targetKey)
-        }
-        hideAdd={pointers.length === 5}
-        centered={true}
-        tabBarGutter={10}
-        items={pointerTabs}
-      />
-    </Card>
-  );
-};
-
-export default GestureEditor;
