+++ app/common/renderer/components/SessionInspector/GesturesTab/GestureEditor/GestureEditor.module.css
